stream_data/consumer_bluesky/requirements.txt
kafka-python
pandas
pyarrow
s3fs
boto3